+++ src/magicdust/constants.js
@@ -1,7 +1,7 @@
 import colors from "./colors";
 
 export default {
-  BORDER_RADIUS: "4px",
+  BORDER_RADIUS: "2px",
   FONT_SIZE: "1.2rem",
   PRIMARY: colors.bar,
   SECONDARY: colors.strokeTwo,